Streams using Dates

+++ JDBCPlayGround.java
@@ -44,7 +44,7 @@ public class JDBCPlayGround {
             stockFundamentalsVO.setTickerSymbol(resultSet.getString("ticker_symbol"));
             stockFundamentalsVO.setSectorId(resultSet.getBigDecimal("sector_id"));
             stockFundamentalsVO.setSubSectorId(resultSet.getInt("subsector_id"));
-            stockFundamentalsVO.setMarketCap(resultSet.getBigDecimal("market_cap"));
+            stockFundamentalsVO.setMarketCap(resultSet.getLong("market_cap"));
             stockFundamentalsVO.setCurrentRation(resultSet.getFloat("current_ratio"));
             specificStockFundamentals.add(stockFundamentalsVO);
         }
